Feature/crew api

+++ src/main/java/com/project/runningcrew/crew/service/CrewService.java
@@ -49,7 +49,10 @@ public class CrewService {
     private final NotificationRepository notificationRepository;
     private final CrewConditionRepository crewConditionRepository;
     private final ImageService imageService;
-    private final String imageDirName = "crew";
+    private final String CREW_IMAGE_DIR_NAME = "crew";
+    private final String DEFAULT_CREW_IMG = "크루 기본 이미지.svg";
+    private final String DEFAULT_CREW_IMG_PATH = CREW_IMAGE_DIR_NAME + "/" + DEFAULT_CREW_IMG;
+
 
     /**
      * 입력받은 crewId 를 가진 Crew 를 찾아 반환한다. 없다면 CrewNotFoundException 을 throw 한다.
@@ -65,18 +68,18 @@ public class CrewService {
     /**
      * 입력받은 Crew 이미지와 Crew 를 저장하고 Crew 의 리더를 생성한 후, Crew 에 부여된 id 를 반환한다.
      *
-     * @param crew          저장할 Crew
-     * @param multipartFile Crew 의 이미지
+     * @param crew 저장할 Crew
      * @return Crew 에 부여된 id
      * @throws CrewNameDuplicateException 입력받은 Crew 의 이름을 가진 크루가 이미 존재할 때
      */
     @Transactional
-    public Long saveCrew(User user, Crew crew, MultipartFile multipartFile) {
+    public Long saveCrew(User user, Crew crew) {
         if (crewRepository.existsByName(crew.getName())) {
             throw new CrewNameDuplicateException(crew.getName());
         }
-        String imageUrl = imageService.uploadImage(multipartFile, imageDirName);
-        crew.updateCrewImgUrl(imageUrl);
+
+        String imgUrl = imageService.getImage(CREW_IMAGE_DIR_NAME, DEFAULT_CREW_IMG);
+        crew.updateCrewImgUrl(imgUrl);
         Crew savedCrew = crewRepository.save(crew);
         crewConditionRepository.save(new CrewCondition(savedCrew));
         memberRepository.save(new Member(user, crew, MemberRole.ROLE_LEADER));
@@ -105,9 +108,12 @@ public class CrewService {
         if (!originCrew.getDongArea().equals(newCrew.getDongArea())) {
             originCrew.updateDongArea(newCrew.getDongArea());
         }
-        if (multipartFile!= null && !multipartFile.isEmpty()) {
-            imageService.deleteImage(originCrew.getCrewImgUrl());
-            String imageUrl = imageService.uploadImage(multipartFile, imageDirName);
+        if (multipartFile != null && !multipartFile.isEmpty()) {
+            String decodeURL = imageService.decodeURL(originCrew.getCrewImgUrl());
+            if (!decodeURL.equals(DEFAULT_CREW_IMG_PATH)) {
+                imageService.deleteImage(originCrew.getCrewImgUrl());
+            }
+            String imageUrl = imageService.uploadImage(multipartFile, CREW_IMAGE_DIR_NAME);
             originCrew.updateCrewImgUrl(imageUrl);
         }
     }
@@ -131,7 +137,10 @@ public class CrewService {
         notificationRepository.deleteAllByCrew(crew);
         crewConditionRepository.deleteByCrew(crew);
         crewRepository.delete(crew);
-        imageService.deleteImage(crew.getCrewImgUrl());
+        String decodeURL = imageService.decodeURL(crew.getCrewImgUrl());
+        if (!decodeURL.equals(DEFAULT_CREW_IMG_PATH)) {
+            imageService.deleteImage(crew.getCrewImgUrl());
+        }
     }
 
     /**
